+++ src/dbi_main.c
@@ -1017,7 +1017,7 @@ void _error_handler(dbi_conn_t *conn, dbi_error_flag errflag) {
 	}
 }
 
-unsigned long _isolate_attrib(unsigned long attribs, unsigned long rangemin, unsigned rangemax) {
+unsigned long _isolate_attrib(unsigned long attribs, unsigned long rangemin, unsigned long rangemax) {
 	/* hahaha! who woulda ever thunk strawberry's code would come in handy? */
 	unsigned short startbit = log(rangemin)/log(2);
 	unsigned short endbit = log(rangemax)/log(2);
